added 2 function (payment gateway stripe, order record result)

+++ src/main/java/com/final_project/e_commerce/service/transaction/TransactionServicerImpl.java
@@ -1,13 +1,16 @@
 package com.final_project.e_commerce.service.transaction;
 
 import com.final_project.e_commerce.data.domainData.reqDomainData.firebaseUser.ReqFirebaseUserDomain;
+import com.final_project.e_commerce.data.domainData.responseDomainData.stripe.ResponseStripeDomain;
 import com.final_project.e_commerce.data.domainData.responseDomainData.transaction.ResponseTransactionDomain;
 import com.final_project.e_commerce.data.entity.cart.CartEntity;
 import com.final_project.e_commerce.data.entity.firebaseUser.FirebaseUserEntity;
 import com.final_project.e_commerce.data.entity.transaction.TransactionEntity;
 import com.final_project.e_commerce.data.entity.transactionProduct.TransactionProductEntity;
+import com.final_project.e_commerce.exception.PaymentNotCompletedException;
 import com.final_project.e_commerce.exception.TransactionIdNotFoundException;
 import com.final_project.e_commerce.exception.TransactionStatusException;
+import com.final_project.e_commerce.mapper.stripe.ChangeToStripeDomain;
 import com.final_project.e_commerce.mapper.transaction.ChangeToDomainTransaction;
 import com.final_project.e_commerce.mapper.transaction.ChangeToTransactionEntity;
 import com.final_project.e_commerce.mapper.transactionProduct.ChangeToTransactionProductEntity;
@@ -16,12 +19,12 @@ import com.final_project.e_commerce.repository.transactionProduct.TransactionPro
 import com.final_project.e_commerce.service.cart.CartService;
 import com.final_project.e_commerce.service.firebaseUser.FirebaseUserService;
 import com.final_project.e_commerce.service.product.ProductService;
+import com.final_project.e_commerce.service.stripe.StripeService;
 import com.final_project.e_commerce.service.transactionProduct.TransactionProductService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import com.stripe.exception.StripeException;
 import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
+import org.springframework.cache.annotation.Caching;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -38,11 +41,12 @@ public class TransactionServicerImpl implements TransactionService {
     private final TransactionProductRepository transactionProductRepository;
     private final ChangeToTransactionProductEntity changeToTransactionProductEntity;
     private final ChangeToDomainTransaction changeToDomainTransaction;
-    private final Logger logger = LoggerFactory.getLogger(TransactionServicerImpl.class);
     private final TransactionProductService transactionProductService;
     private final ProductService  productService;
+    private final StripeService stripeService;
+    private final ChangeToStripeDomain changeToStripeDomain;
 
-    public TransactionServicerImpl(FirebaseUserService firebaseUserService, CartService cartService, ChangeToTransactionEntity changeToTransactionEntity, TransactionRepository transactionRepository, TransactionProductRepository transactionProductRepository, ChangeToTransactionProductEntity changeToTransactionProductEntity, ChangeToDomainTransaction changeToDomainTransaction, TransactionProductService transactionProductService, ProductService productService) {
+    public TransactionServicerImpl(FirebaseUserService firebaseUserService, CartService cartService, ChangeToTransactionEntity changeToTransactionEntity, TransactionRepository transactionRepository, TransactionProductRepository transactionProductRepository, ChangeToTransactionProductEntity changeToTransactionProductEntity, ChangeToDomainTransaction changeToDomainTransaction, TransactionProductService transactionProductService, ProductService productService, StripeService stripeService, ChangeToStripeDomain changeToStripeDomain) {
         this.firebaseUserService = firebaseUserService;
         this.cartService = cartService;
         this.changeToTransactionEntity = changeToTransactionEntity;
@@ -52,6 +56,8 @@ public class TransactionServicerImpl implements TransactionService {
         this.changeToDomainTransaction = changeToDomainTransaction;
         this.transactionProductService = transactionProductService;
         this.productService = productService;
+        this.stripeService = stripeService;
+        this.changeToStripeDomain = changeToStripeDomain;
     }
 
 
@@ -83,7 +89,7 @@ public class TransactionServicerImpl implements TransactionService {
     @CacheEvict(cacheNames = "getTransaction", key = "#tid")
     @Transactional
     @Override
-    public void updateTransactionStatusToProcessing(ReqFirebaseUserDomain reqFirebaseUserDomain, Integer tid) {
+    public ResponseStripeDomain updateTransactionStatusToProcessing(ReqFirebaseUserDomain reqFirebaseUserDomain, Integer tid) throws StripeException {
         FirebaseUserEntity firebaseUserEntity = firebaseUserService.getFirebaseUserByEmail(reqFirebaseUserDomain);
         Optional<TransactionEntity> transactionByTid = transactionRepository.getTransactionByTid(firebaseUserEntity.getUid(), tid);
         if (transactionByTid.isEmpty()) {
@@ -94,12 +100,16 @@ public class TransactionServicerImpl implements TransactionService {
             throw new TransactionStatusException(tid, "PREPARE");
         }
         transactionEntity.setStatus(TransactionStatusEnum.PROCESSING);
+        return changeToStripeDomain.changeToStripeDomain(stripeService.creatStripePaymentSession(transactionEntity));
     }
 
-    @CacheEvict(cacheNames = "getTransaction", key = "#tid")
+    @Caching(evict = {
+            @CacheEvict(cacheNames = "getTransaction", key = "#tid"),
+            @CacheEvict(cacheNames = "getUserOrderRecordList", key = "#reqFirebaseUserDomain.email"),
+    } )
     @Transactional
     @Override
-    public ResponseTransactionDomain updateTransactionStatusToSuccess(ReqFirebaseUserDomain reqFirebaseUserDomain, Integer tid) {
+    public ResponseTransactionDomain updateTransactionStatusToSuccess(ReqFirebaseUserDomain reqFirebaseUserDomain, Integer tid) throws StripeException {
         FirebaseUserEntity firebaseUserEntity = firebaseUserService.getFirebaseUserByEmail(reqFirebaseUserDomain);
         Optional<TransactionEntity> transactionByTid = transactionRepository.getTransactionByTid(firebaseUserEntity.getUid(), tid);
         if (transactionByTid.isEmpty()) {
@@ -109,11 +119,22 @@ public class TransactionServicerImpl implements TransactionService {
         if (transactionEntity.getStatus() != TransactionStatusEnum.PROCESSING){
             throw new TransactionStatusException(tid, "PROCESSING");
         }
+        if (!stripeService.isPaymentCompleted(transactionEntity)){
+            throw new PaymentNotCompletedException(tid);
+        }
         List<CartEntity> cartItemEntityList = cartService.getCartItemEntityListByUid(firebaseUserEntity);
         productService.deduceStock(cartItemEntityList);
         transactionEntity.setStatus(TransactionStatusEnum.SUCCESS);
         cartService.deleteCartItemByUserId(firebaseUserEntity);
         List<TransactionProductEntity> transactionProductList = transactionProductService.getTransactionProductByTid(tid);
         return changeToDomainTransaction.transactionEntityChangeToResponseTransactionDomain(transactionEntity, transactionProductList);
+    }
+
+    @Cacheable(cacheNames = "getUserOrderRecordList", key = "#reqFirebaseUserDomain.email")
+    @Override
+    public List<ResponseTransactionDomain> getUserOrderRecordList(ReqFirebaseUserDomain reqFirebaseUserDomain){
+        FirebaseUserEntity firebaseUserEntity = firebaseUserService.getFirebaseUserByEmail(reqFirebaseUserDomain);
+        List<TransactionEntity> transactionEntityList = transactionRepository.getUserOrderRecordList(firebaseUserEntity.getUid(), TransactionStatusEnum.SUCCESS.name());
+        return changeToDomainTransaction.transactionEntityListToResponseTransactionDomain(transactionEntityList);
     }
 }
